// src/types.ts
export type Service = {
  title: string;
  description: string;
  icon: string;
};

export type BlogPost = {
  id: number;
  title: string;
  excerpt: string;
  date: string;
  image: string;
  slug: string;
};

export type WhyChooseItem = {
  title: string;
  description: string;
  icon: string;
};

export type HowStep = {
  title: string;
  description: string;
};